Rework the support for finding APR/APU paths.
The previous method for finding APR/APU paths didn't really work. It
just seemed to locally.

The new approach is to look for a few known, hopefully stable paths. It
now looks for the Homebrew apr-1-config and apu-1-config binaries, and
uses them to look for directories to search for include files and
libraries. It also considers the versions in the XCode directories.

If we find suitable paths, we'll prioritize those. Otherwise, we let
PySVN use its own, more outdated path scanning.