Fix the configuration of the APU include directory.
When attempting to determine a suitable APU include directory, we were
looking for an `apr.h` file. This wasn't quite correct -- we really
wanted `apu.h`. This meant that if you didn't have a suitable system
version, it would outright fail.

This fixes up the check, looking for versions installed by XCode and
homebrew.