Look up Subversion paths provided by brew on macOS.
This enhances PySVN on macOS by directly supporting brew's installs of
Subversion. When installed via brew, PySVN will be given the location to
the bin, lib, and include directories, ensuring that those are used
instead of the system install.

This is important as newer versions of XCode no longer appear to include
the header files themselves, so this provides a path to install it
seamlessly on macOS.