Add support for compiling against a Homebrew install on Apple M1.
On M1 CPUs, Homebrew installs into /opt/homebrew, rather than
/usr/local. We had some hard-coded paths that searched /usr/local,
causing Homebrew versions of Subversion to be ignored.

We now call `brew --prefix` for the packages, rather than hard-coding
paths, ensuring that we'll get the right place now for Intel or M1
processors, or if Homebrew ever changes paths in the future.