Add support for installing on Bitnami.
Bitnami installs are self-contained, with their own versions of the
Subversion and Apache libraries, and this can make it tricky to install
on them.

When installing on Linux as the "bitnami" user, the script now
hard-codes the paths for the Bitnami install rather than using the
system install.

This repurposes some of the brew code we had for apr-1-config and
apu-1-config, making it available in the Bitnami case as well. We should
also be using this for Linux more generally, but that's out of scope for
this work.